separated ExpandableBlock into subcomponents
